# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Grades API",
        "description": "API Documentation for Student Grade and Curriculum Management",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8003",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/api/v1/curriculums": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Daftar aturan prasyarat kurikulum",
                "description": "Menampilkan daftar aturan prasyarat kurikulum program studi untuk mendeteksi keterikatan antar mata kuliah",
                "operationId": "580dfd2db9d016e66014796821f3c4e9",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "course_code": {
                                                        "type": "string",
                                                        "example": "IF101"
                                                    },
                                                    "course_name": {
                                                        "type": "string",
                                                        "example": "Dasar Pemrograman"
                                                    },
                                                    "prerequisite": {
                                                        "type": "string",
                                                        "example": "-"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Grades-and-Curriculum-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/grades": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Menampilkan semua data nilai mahasiswa",
                "description": "Menampilkan daftar seluruh data nilai mahasiswa",
                "operationId": "b8486dec7450301f5b89d58208649f9e",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "student_id": {
                                                        "type": "string",
                                                        "example": "102022400285"
                                                    },
                                                    "course_code": {
                                                        "type": "string",
                                                        "example": "SI4808"
                                                    },
                                                    "grade": {
                                                        "type": "string",
                                                        "example": "A"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "LULUS"
                                                    },
                                                    "receipt_number": {
                                                        "type": "string",
                                                        "example": "REC-12345"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Grades-Curriculum-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Initialize student grade record (Critical Transaction)",
                "operationId": "19f4ae3304228b18e0f09491e2a276ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "student_id": {
                                        "type": "string",
                                        "example": "102022400285"
                                    },
                                    "course_id": {
                                        "type": "string",
                                        "example": "SI4808"
                                    },
                                    "course_code": {
                                        "type": "string",
                                        "example": "SI4808"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Grade initialized and audited successfully"
                    },
                    "401": {
                        "description": "Invalid API Key"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/grades/{student_id}": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Menampilkan detail riwayat transkrip nilai mahasiswa",
                "description": "Menampilkan detail riwayat transkrip nilai mahasiswa untuk pembuktian kelulusan mata kuliah prasyarat",
                "operationId": "a92907dad8768fdfb5dfc102dcdf5d52",
                "parameters": [
                    {
                        "name": "student_id",
                        "in": "path",
                        "description": "NIM Mahasiswa",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grade records found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "student_id": {
                                                        "type": "string",
                                                        "example": "102022400285"
                                                    },
                                                    "course_code": {
                                                        "type": "string",
                                                        "example": "SI4808"
                                                    },
                                                    "grade": {
                                                        "type": "string",
                                                        "example": "A"
                                                    },
                                                    "status": {
                                                        "type": "string",
                                                        "example": "LULUS"
                                                    },
                                                    "receipt_number": {
                                                        "type": "string",
                                                        "example": "REC-12345"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "service_name": {
                                                    "type": "string",
                                                    "example": "Grades-Curriculum-Service"
                                                },
                                                "api_version": {
                                                    "type": "string",
                                                    "example": "v1"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Grade records not found"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Grades",
            "description": "Grades & Curriculum API"
        }
    ]
}